:recycle: watchdog: refactor logging and constants
